Add test for ignore_participant
